+++ src/index.js
@@ -3,4 +3,4 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 export { renderAretino } from './renderer.js';
-export { parseAretino } from './parser.js';
+export { parseAretino, matchAccidental } from './parser.js';